+++ packages/lesmis/src/utils/export/exportImagesToPDF.js
