+++ packages/lesmis/src/utils/export/getImages/useCustomGetImages.js
@@ -3,15 +3,15 @@
  *  Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
  */
 
-import { useRef, useState } from 'react';
+import { useRef } from 'react';
 
 import { getImage } from './getImage';
 
 // Customised for dashboards pdf export
 export const useCustomGetImages = () => {
-  const [isExporting, setIsExporting] = useState(false);
   const refs = useRef([]);
   refs.current = [];
+
   const addToRefs = el => {
     if (el && !refs.current.includes(el)) {
       refs.current.push(el);
@@ -19,9 +19,7 @@ export const useCustomGetImages = () => {
   };
 
   const getImgs = async () => {
-    setIsExporting(true);
     const imgs = [];
-
     for (let i = 0; i < refs.current.length; i++) {
       // Hacky way to take screenshot for the first profile page with `html2canvas` as `domtoimage` can't access material-ui css file.
       // More on this, `html2canvas` can return a better resolution screenshots than `domtoimage`, but it requires much longer time.
@@ -30,9 +28,8 @@ export const useCustomGetImages = () => {
       imgs.push(image);
     }
 
-    setIsExporting(false);
     return imgs;
   };
 
-  return { isExporting, addToRefs, getImgs };
+  return { addToRefs, getImgs };
 };
