+++ packages/lesmis/src/utils/export/index.js
@@ -1,0 +1,3 @@
+export { exportImagesToPDF } from './exportImagesToPDF';
+export { useExportToImage } from './useExportToImage';
+export * from './getImages';
