+++ packages/lesmis/src/utils/index.js
@@ -13,5 +13,4 @@ export * from './useStickyBar';
 export * from './useUrlParams';
 export * from './useUrlSearchParams';
 export * from './useEntityBreadcrumbs';
-export * from './useImages';
-export * from './exportImagesToPDF';
+export * from './export';
